pyproject.toml
[project]
name = "announcement-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "penne",email = "penne0505pp@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py (>=2.6.4,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "asyncpg (>=0.29.0,<0.31.0)",
]

[project.scripts]
announcement-bot = "app.runtime:main"

[tool.poetry]
packages = [
    { include = "app", from = "src" },
    { include = "bot", from = "src" },
    { include = "views", from = "src" },
]

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = ">=8.2,<9.0"
pytest-asyncio = ">=0.25,<0.26"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
